requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.7.3
aiosignal==1.3.2
APScheduler==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
click==7.1.2
constants==0.6.0
decorator==5.1.1
Flask==1.1.2
frozenlist==1.5.0
gunicorn==20.0.4
idna==2.10
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multi_key_dict==2.0.3
multidict==5.1.0
paramiko==2.7.2
pathtools==0.1.2
pbr==5.5.1
propcache==0.2.1
pycparser==2.20
pylint==2.6.0
PyNaCl==1.4.0
python-dotenv==0.15.0
python-jenkins==1.8.2
pytz==2020.5
requests==2.24.0
simplejson==3.17.2
six==1.15.0
slackclient==2.9.3
toml==0.10.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
validators==0.34.0
watchdog==0.10.4
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.18.3
